Indev Update 8
This one took a while, but I've now laid the groundwork for all that's needed to begin creating the NPC system and merchants. For now, there's little new content as a lot of changes that were made are not obvious in game.
This update drastically changes how the core game works, since there's no more RNG when it comes to your equipment. Balancing is obviously still underway!